Move handling of deep link clicks to App.tsx
